Articulo y ArticuloDTO: Se cambia int a long en Codigo

+++ Tienda.LogicaAplicacion/DTOs/ArticuloDTO.cs
@@ -11,7 +11,7 @@ namespace Tienda.LogicaAplicacion.DTOs
     {
         public int Id { get; set; }
         public string Nombre { get; set; }
-        public int Codigo { get; set; }
+        public long Codigo { get; set; }
         public string Descripcion { get; set; }
         public int stock { get; set; }
         public decimal PrecioUnitario { get; set; }
